requirements.txt
langchain>=0.0.200
openai>=0.27.0
pydantic>=2.0.0
python-dotenv>=0.19.0
pytest>=7.0.0
pytest-asyncio>=0.20.0
typing-extensions>=4.0.0

